from_and_to_copydown.cpp
This code copies down the from and to stops in route atlas where from and to stops field is blank. It first checks the route number and if that route number has from and to stops then it copies it down.For this code to work route number should be your first column in your file and make sure that you already copied down the route number.