uptrip_stop_times.cpp
This code generate a file containing all details required to make trip id,stop sequence of that trip with their stop codes and arrival and departure times for first and last stop. This is only for the trips that are from (from) stop to (to) stop in route atlas.It takes input atlas file which only contain uptrips i.e. only the trips from (from) to (to) stops of atlas.Uptrip contains two columns which have the stop serial number of the from and to stops of atlas as they are in route details.