downtrip_stop_times.cpp
This code generate a file containing all details required to make trip id,stop sequence of that trip with their stop codes and arrival and departure times for first and last stop. This is only for the trips that are from (to) stop to (from) stop in route atlas.It takes input atlas file which only contain downtrips i.e. only the trips from (to) to (from) stops of atlas.Downtrip contains two columns which have the stop serial number of the to and from stops of atlas as they are in route details.Make sure that the starting stop serial number is now of the to stop and ending stop serial number is of the from stop.